+++ game_engine/targets/windows/core/include/Rendering/ShaderBackend.hpp
@@ -70,7 +70,7 @@ public:
                 {"TANGENT", 0, DXGI_FORMAT_R32G32B32_FLOAT, 0, D3D11_APPEND_ALIGNED_ELEMENT, D3D11_INPUT_PER_VERTEX_DATA, 0}};
 
             input_layout_.reset(new InputLayout(
-                *gfx, ied, std::size(ied),
+                *gfx, ied, static_cast<UINT>(std::size(ied)),
                 sub_shaders_[0].vertex_shader->bytecode().GetBufferPointer(),
                 sub_shaders_[0].vertex_shader->bytecode().GetBufferSize()));
         }
